[SPIRV] byte[] vectors memory access fixed

+++ unittests/src/main/java/uk/ac/manchester/tornado/unittests/arrays/TestArrays.java
@@ -89,7 +89,7 @@ public class TestArrays extends TornadoTestBase {
 
     public static void initializeSequentialByte(byte[] a) {
         for (int i = 0; i < a.length; i++) {
-            a[i] = 21;
+            a[i] = (byte) 21;
         }
     }
